+++ packages/components/src/tools-panel/tools-panel-header/component.tsx
@@ -1,6 +1,7 @@
 /**
  * External dependencies
  */
+import clsx from 'clsx';
 import type { ForwardedRef } from 'react';
 
 /**
@@ -8,7 +9,7 @@ import type { ForwardedRef } from 'react';
  */
 import { speak } from '@wordpress/a11y';
 import { check, moreVertical, plus } from '@wordpress/icons';
-import { __, _x, sprintf } from '@wordpress/i18n';
+import { __, _x, sprintf, isRTL } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
@@ -19,9 +20,9 @@ import MenuItem from '../../menu-item';
 import { HStack } from '../../h-stack';
 import { Heading } from '../../heading';
 import { useToolsPanelHeader } from './hook';
+import styles from '../style.module.scss';
 import type { WordPressComponentProps } from '../../context';
 import { contextConnect } from '../../context';
-import { ResetLabel } from '../styles';
 import type {
 	ToolsPanelControlsGroupProps,
 	ToolsPanelHeaderProps,
@@ -36,7 +37,17 @@ const DefaultControlsGroup = ( {
 		return null;
 	}
 
-	const resetSuffix = <ResetLabel aria-hidden>{ __( 'Reset' ) }</ResetLabel>;
+	const resetSuffix = (
+		<span
+			aria-hidden
+			className={ clsx(
+				styles[ 'reset-label' ],
+				isRTL() && styles[ 'reset-label-rtl' ]
+			) }
+		>
+			{ __( 'Reset' ) }
+		</span>
+	);
 
 	return (
 		<>